+++ Tests/Unit/Module/AuditPolicyRuleAdvanced.tests.ps1
@@ -1,0 +1,38 @@
+#region Header
+. $PSScriptRoot\.tests.header.ps1
+#endregion
+
+try
+{
+    InModuleScope -ModuleName "$($global:moduleName).Convert" {
+        #region Test Setup
+        $testRuleList = @(
+            @{
+                Subcategory                 = 'Process Creation'
+                AuditFlag                   = 'Failure'
+                Ensure                      = 'Present'
+                OrganizationValueRequired   = $false
+                OrganizationValueTestString = $null
+                CheckContent                = 'Ensure Audit Process Creation auditing has been enabled:
+
+Computer Configuration >> Windows Settings >> Security Settings >> Advanced Audit Policy Configuration >> System Audit Policy >> Detailed Tracking >> Audit Process Creation.
+
+If "Audit Process Creation" is not set to "Failure", this is a finding.'
+            }
+        )
+        #endregion
+
+        foreach ($testRule in $testRuleList)
+        {
+            . $PSScriptRoot\Convert.CommonTests.ps1
+        }
+
+        #region Add Custom Tests Here
+
+        #endregion
+    }
+}
+finally
+{
+    . $PSScriptRoot\.tests.footer.ps1
+}